refactor for adding different components
