+++ examples/workflow.py
@@ -2,15 +2,15 @@ from indico import IndicoClient, IndicoConfig
 from indico.queries import (
     JobStatus,
     RetrieveStorageObject,
-    WorkflowSubmission, CreateWorkflow, GetDataset, AddModelGroupComponent, ListWorkflows,
+    WorkflowSubmission, CreateWorkflow, GetDataset, ListWorkflows, AddModelGroupComponent,
 )
 
 # Use your dataset's id to call its associated workflow
 dataset_id = 6826
 workflow_name = "Test Workflow"
-#the name of the datacolumn in your dataset
+# the name of the datacolumn in your dataset
 datacolumn = "column_name"
-#name of an existing labelset.
+# name of an existing labelset.
 labelset_name = "labelset_name"
 
 my_config = IndicoConfig(
@@ -20,15 +20,16 @@ client = IndicoClient(config=my_config)
 
 dataset = client.call(GetDataset(id=dataset_id))
 
-#Example 1: Create a new workflow and model on an existing dataset.
-#create a new workflow
+# Example 1: Create a new workflow and model on an existing dataset.
+# create a new workflow
 new_workflow = client.call(CreateWorkflow(name=workflow_name, dataset_id=dataset_id))
 
-#First, find the component you want this model to follow.
-#in this case, we want it after OCR.
+# First, find the component you want this model to follow after.
+# A component is essentially a step in the workflow process.
+# in this case, we want it after OCR.
 after_component_id = new_workflow.component_by_type("INPUT_OCR_EXTRACTION").id
 
-#now, create the request with your parameters.
+# now, create the request with your parameters.
 modelgroupreq = AddModelGroupComponent(
     name=workflow_name,
     dataset_id=dataset.id,
@@ -40,12 +41,12 @@ modelgroupreq = AddModelGroupComponent(
 
 client.call(modelgroupreq)
 
-#Example 2: Simple submission.
+# Example 2: Simple submission.
 
 # Return a list of workflows for this dataset id or an empty list if there are none
 workflows = client.call(ListWorkflows(dataset_ids=[dataset_id]))
 
-#You can run this with the newly created workflow or with an existing one.
+# You can run this with the newly created workflow or with an existing one.
 if workflows:
     # Send a document through the workflow
     # Get back one Job per file
