+++ indico/queries/__init__.py
@@ -7,3 +7,4 @@ from .submission import *
 from .workflow import *
 from .forms import *
 from .export import *
+from .workflow_components import *
