+++ indico/queries/model_groups/model_groups.py
@@ -5,8 +5,10 @@ from typing import List, Dict
 import deprecation
 
 from indico.client.request import GraphQLRequest, RequestChain
+from indico.queries.workflow_components import AddModelGroupComponent
 from indico.types import Workflow
-from indico.types.model_group import ModelGroup, NewQuestionaireArguments, NewLabelsetArguments
+from indico.types.model_group import ModelGroup, NewLabelsetArguments, \
+    NewQuestionnaireArguments
 from indico.types.model import Model
 from indico.types.jobs import Job
 from indico.types.utils import cc_to_snake
@@ -242,7 +244,7 @@ class CreateModelGroup(RequestChain):
                 sleep(1)
                 yield req
 
-            yield _GetModelGroup(id=model_group_id)
+        yield _GetModelGroup(id=model_group_id)
 
 
 class LoadModel(GraphQLRequest):
@@ -361,125 +363,3 @@ class ModelGroupPredict(RequestChain):
             model_id=self.model_id, data=self.data, predict_options=self.predict_options
         )
 
-
-class AddModelGroupComponent(GraphQLRequest):
-    """
-    Adds a new model group component to a workflow, optionally with a customized questionnaire. Available on 5.0+ only.
-    Returns workflow with updated component list.
-    Args:
-         workflow_id(int),
-         dataset_id(int),
-         name(str),
-        source_column_id(str),
-        after_component_id(str),
-        labelset_column_id(int),
-        new_labelset_args: NewLabelsetArguments = None
-        new_questionnaire_args: NewQuestionaireArguments = None
-
-    """
-    query = """
-            mutation addModelGroup(
-          $workflowId: Int!, 
-          $name: String!, 
-          $datasetId: Int!, 
-          $sourceColumnId: Int!, 
-          $afterComponentId: Int, 
-          $labelsetColumnId: Int,
-          $newLabelsetArgs: NewLabelsetInput,
-          $questionnaireArgs: QuestionnaireInput,
-          $modelTrainingOptions: JSONString,
-          $modelType : ModelType
-        ) {
-          addModelGroupComponent(workflowId: $workflowId, name: $name, datasetId: $datasetId, 
-          sourceColumnId: $sourceColumnId, afterComponentId: $afterComponentId, labelsetColumnId: $labelsetColumnId,
-          modelTrainingOptions: $modelTrainingOptions,
-
-    newLabelsetArgs: $newLabelsetArgs,
-    questionnaireArgs: $questionnaireArgs, modelType: $modelType) {
-            workflow {
-                id
-                components {
-                                id
-                                componentType
-                                reviewable
-                                
-                                filteredClasses
-                                ... on ModelGroupComponent {
-                                    taskType
-                                    modelType
-                                    modelGroup {
-                                        status
-                                      id
-                                      name
-                                      taskType
-                                      questionnaireId
-                                      selectedModel{
-                                        id
-                                      }
-                                    }
-                                }
-
-                            }
-                            componentLinks {
-                                id
-                                headComponentId
-                                tailComponentId
-
-                            }
-
-            }
-          }
-        }
-            """
-
-    def __init__(self, workflow_id: int, dataset_id: int, name: str,
-                 source_column_id: int, after_component_id: int = None, labelset_column_id: int = None,
-                 new_labelset_args: NewLabelsetArguments = None,
-                 new_questionnaire_args: NewQuestionaireArguments = None, model_training_options: str = None,
-                 model_type: str = None):
-        if labelset_column_id is not None and new_labelset_args is not None:
-            raise IndicoInputError("Cannot define both labelset_column_id and new_labelset_args, must be one "
-                                   "or the other.")
-        if labelset_column_id is None and new_labelset_args is None:
-            raise IndicoInputError("Must define one of either labelset_column_id or new_labelset_args.")
-
-        super().__init__(
-            self.query,
-            variables={
-                "workflowId": workflow_id,
-                "name": name,
-                "datasetId": dataset_id,
-                "sourceColumnId": source_column_id,
-                "labelsetColumnId": labelset_column_id,
-                "afterComponentId": after_component_id,
-                "modelTrainingOptions": model_training_options,
-                "modelType": model_type,
-                "newLabelsetArgs": self.__labelset_to_json(
-                    new_labelset_args) if new_labelset_args is not None else None,
-                "questionnaireArgs": self.__questionnaire_to_json(
-                    new_questionnaire_args) if new_questionnaire_args is not None else None
-
-            }
-        )
-
-    def __labelset_to_json(self, labelset: NewLabelsetArguments):
-        targets = ', '.join(f'"{w}"' for w in labelset.target_names)
-        return {
-            "name": labelset.name,
-            "numLabelersRequired": labelset.num_labelers_required,
-            "datacolumnId": labelset.datacolumn_id,
-            "taskType": labelset.task_type,
-            "targetNames": labelset.target_names
-        }
-
-    def __questionnaire_to_json(self, questionnaire: NewQuestionaireArguments):
-        return {
-            "instructions": questionnaire.instructions,
-            "forceTextMode": questionnaire.force_text_mode,
-            "showPredictions": questionnaire.show_predictions,
-            "users": questionnaire.users
-
-        }
-
-    def process_response(self, response) -> Workflow:
-        return Workflow(**super().process_response(response)["addModelGroupComponent"]["workflow"])
